renamed test_append to check_append to exclude the script from the action tests
