+++ src/itertree/examples/performance_analysis/check_compare.py
