+++ src/itertree/examples/performance_analysis/check_copy.py
