+++ src/itertree/examples/performance_analysis/check_delete.py
