+++ src/itertree/examples/performance_analysis/check_extend.py
